Moving to default adding comment, cleanup
